+++ src/generic_back/track/audio_track.rs
@@ -23,7 +23,7 @@ impl AudioTrack {
         }))
     }
 
-    pub fn fill_buf(&self, buf_start_sample: u32, buf: &mut [f32]) {
+    pub fn fill_buf(&self, buf_start_sample: usize, buf: &mut [f32]) {
         if !self.meter.playing.load(SeqCst) && !self.meter.exporting.load(SeqCst) {
             return;
         }
